spotyxbmc: Add ramdisk feature for spotify cache. Good for audio hickups when using slow disks.
When enabled, it does cause slower spotify startup because
cache contents gets copied to RAM before we can start spotify.
Changes to the ramdisk is sent to disk every two seconds.

When ramdisk is enabled it is suggested to perform reboot or
shutdown from within XBMC (compared to reboot/shutdown using ssh)